added forgot password functionality

+++ src/api/resolvers/post/mutations.ts
@@ -1,6 +1,9 @@
+import { ObjectId } from "mongoose";
 import { ICreatePostInput, IPost, IToggleLikeInput } from "../../../interfaces/IPost";
+import { IUser } from "../../../interfaces/IUser";
 import { deletePostById } from "../../../services/posts";
 import createPost from "../../../services/posts/createPost";
+import { getPost } from "../../../services/posts/getPost";
 import toggleLike from "../../../services/posts/toggleLike";
 
 const resolver = {
@@ -9,22 +12,45 @@ const resolver = {
      *
      * @param { void } _parent Previous resolver.
      * @param { string } postId The post id.
+     * @param { { IUser } } context Authenticated user.
      * @returns { Promise<IPost> } The post, false value if not found.
      */
-    deletePost: async (_parent: void, { postId }: { postId: string }): Promise<IPost> => {
-        const post = await deletePostById(postId);
-        return post;
+    deletePost: async (_parent: void, { postId }: { postId: string }, { user }: { user: IUser }): Promise<IPost> => {
+        if (!user) {
+            throw new Error("This option is valid only for authenticated users.");
+        }
+
+        const post = await getPost(postId);
+
+        if(post.author.id !== user.id) {
+            throw new Error("Only the author of this post can delete it.");
+        }
+
+        const deletedPost = await deletePostById(postId);
+        return deletedPost;
     },
 
     /**
      * Creates a post.
      *
      * @param { void } _parent Previous resolver.
-     * @param { ICreatePostInput } postInput The post to create.
+     * @param { { title: string, body: string } } postInput The post to create.
+     * @param { { IUser } } context Authenticated user.
      * @returns { Promise<IPost> } The post that just created.
      */
-    createPost: async (_parent: void, { createPostInput }: { createPostInput: ICreatePostInput }): Promise<IPost> => {
-        const post = await createPost(createPostInput);
+    createPost: async (_parent: void, { title, body }: { title: string, body: string }, { user }: { user: IUser }): Promise<IPost> => {
+        if (!user) {
+            throw new Error("This option is valid only for authenticated users.");
+        }
+
+        const postInput: ICreatePostInput = {
+            author: user.id,
+            body,
+            title,
+            likes: [],
+        };
+
+        const post = await createPost(postInput);
         return post;
     },
 
@@ -32,11 +58,21 @@ const resolver = {
      * Toggles like/dislike for a post from a given user.
      *
      * @param { void } _parent Previous resolver.
-     * @param { ICreatePostInput } toggleLikeInput The user and the post to be liked.
+     * @param { ObjectId } postId The post to be liked.
+     * @param { { IUser } } context Authenticated user.
      * @returns { Promise<IPost> } The post liked.
      */
-    toggleLike: async (_parent: void, { toggleLikeInput }: { toggleLikeInput: IToggleLikeInput }): Promise<IPost> => {
-        const post = await toggleLike(toggleLikeInput);
+    toggleLike: async (_parent: void, { postId }: { postId: ObjectId }, { user }: { user: IUser }): Promise<IPost> => {
+        if (!user) {
+            throw new Error("This option is valid only for authenticated users.");
+        }
+
+        const likeInput: IToggleLikeInput = {
+            postId,
+            userId: user.id,
+        };
+
+        const post = await toggleLike(likeInput);
         return post;
     },
 };
